Patch: none path handling;

+++ mypass/db/tiny/_dao.py
@@ -1,35 +1,39 @@
 import abc
 import os
 from pathlib import Path
+from typing import Type
 
 from tinydb import TinyDB, Storage
 
-DB_PATH = Path.home().joinpath('.mypass', 'db', 'tinydb', 'db.json')
-
 
 class TinyDao(abc.ABC):
-    def __init__(self, path: str | os.PathLike = None, storage: Storage = None, *args, **kwargs):
-        if path is None:
-            path = DB_PATH
-        path = Path(path)
+    def __init__(self, path: str | os.PathLike = None, storage: Type[Storage] = None, *args, **kwargs):
+        if path is not None:
+            path = Path(path)
         self._path = path
         self._storage = storage
         self._storage_args = args
         self._storage_kwargs = kwargs
         if storage is not None:
             self._storage_kwargs['storage'] = storage
+        if path is not None:
+            self._storage_kwargs['path'] = path
 
     def init_db(self):
+        if self._path is None:
+            return
+
         if not self._path.parent.exists():
             self._path.parent.mkdir(parents=True)
             self._path.touch()
 
     def get_connection(self):
         self.init_db()
-        return TinyDB(self._path, *self._storage_args, **self._storage_kwargs)
+        return TinyDB(*self._storage_args, **self._storage_kwargs)
 
     def unlink(self):
-        unlink(self._path)
+        if self._path is not None:
+            unlink(self._path)
 
     @abc.abstractmethod
     def create(self, *args, **kwargs):
@@ -56,9 +60,7 @@ class TinyDao(abc.ABC):
         ...
 
 
-def unlink(path: str | os.PathLike = None):
-    if path is None:
-        path = DB_PATH
+def unlink(path: str | os.PathLike):
     path = Path(path)
 
     if path.exists():
